fix: set pyproject.toml python version to >=3.10
the project was set to run on python 3.10.8 in the huggingface
spaces but the pyproject.toml was requiring an newer version.

while this does not affect the build process or how it runs on HF,
without that fixed value uv would not allow you to run `uv run` to
do local development.